Attempt to solve 404 bug when site is deployed to GitHub pages
When deploying this site to Github pages, the root page "/" works as expected, but when I go to /about or /blog, it *only* works when clicking the link in the header @src\components\Header.tsx. Whenever I refresh the page or try to go directly there, i.e. go to https://oscarbennich.github.io/about in the browser, I get a 404 error.